+++ spec/integration_helper.rb
@@ -79,7 +79,7 @@ RSpec.configure do |config|
   config.mock_with :rspec
   config.expect_with :rspec
   config.treat_symbols_as_metadata_keys_with_true_values = true
-  config.filter_run_excluding :integration => true
+  config.filter_run_excluding :integration => true unless ENV['ONE']
 
   # Create the Peas container against which the CLI client will interact
   config.before(:all, :integration) do